tests/cpu_esp: test for heap using esp_idf_heap and esp_spi_ram
The test is just a copy of tests/malloc but uses the heap of the SDK (module esp_idf_heap) instead of the heap provided by the newlib. If feature esp_spi_ram is provided, the module esp_esp_ram is enabled. Thus the test is good for testing modules esp_idf_heap and esp_spi_ram.